Poet-page search now matches poem title + content (poet-scoped)
- /api/search accepts an optional ?poet=<id> scope (validated ObjectId)
- Extract shared PoemSearchBox (pill + debounced API + motion results +
  highlight + snippets); poet-scoped results hide the redundant poet name
- HomeSearch becomes a thin section wrapper (now a server component);
  PoemSearch is a thin poet-scoped wrapper passing poet.id
